Chrome Engine 6 needs explicit dds extension added manually
CEngine stores the original file name as the base name
instead of Texture it stores Texture.dds or Texture.png or Texture.raw

While internally all of these are still dds, I (should) convert them back to their original format, and I do so with png at least.

although it means that until now I didn't need to add the .dds extension as it was already there

but now I support CE6, so I need to actually add that now.